requirements.txt
pytz==2024.1
pyreadline3==3.4.1
tqdm==4.66.2
flask_cors===4.0.0
mysql-connector-python==8.3.0
pymysql==1.1.0
dbutils==3.1.0
tortoise-orm==0.20.0
asyncmy==0.2.9
aerich==0.7.2
fastapi==0.110.3
python-multipart==0.0.9
uvicorn[standard]==0.29.0
pytest==8.2.0
asgi_lifespan==2.1.0
httpx==0.27.0
bson==0.5.10
psutil==5.9.8
bcrypt==4.1.2
